separate missions by types; change launch file to include launch per type of mission; change gui to show only one type of mission
