+++ gui/src/gui/choose_gazebo.py
@@ -1,0 +1,145 @@
+import os
+import rospy
+import QtGui
+from qt_gui.plugin import Plugin
+from python_qt_binding import loadUi
+from python_qt_binding.QtGui import QWidget
+from PyQt4.QtCore import QObject, pyqtSignal
+import PyQt4.QtCore
+import pyqtgraph as pg
+
+# necessary to have gui as a client, asking controller to save data
+# from python_qt_binding.QtCore import QTimer, Slot
+# from python_qt_binding.QtCore import pyqtSlot
+
+
+import numpy
+
+# import analysis
+# import utils
+import subprocess
+
+
+
+# import services defined in quad_control
+# SERVICE BEING USED: Simulator_srv
+from quad_control.srv import *
+
+
+import argparse
+
+
+import rospkg
+# get an instance of RosPack with the default search paths
+rospack = rospkg.RosPack()
+# get the file path for rospy_tutorials
+import sys
+sys.path.insert(0, rospack.get_path('quad_control'))
+# no need to get quad_control path, since it is package; import controllers dictionary
+from src.simulators import simulators_dictionary
+
+
+from choose_jsonable import ChooseJsonablePlugin
+
+class ChooseGazeboPlugin(Plugin):
+
+
+    def __init__(self, context,namespace = None):
+
+        # it is either "" or the input given at creation of plugin
+        self.namespace = self._parse_args(context.argv())
+
+
+        super(ChooseGazeboPlugin, self).__init__(context)
+        # Give QObjects reasonable names
+        self.setObjectName('ChooseGazeboPlugin')
+
+        # Process standalone plugin command-line arguments
+        from argparse import ArgumentParser
+        parser = ArgumentParser()
+        # Add argument(s) to the parser.
+        parser.add_argument("-q", "--quiet", action="store_true",
+                      dest="quiet",
+                      help="Put plugin in silent mode")
+        args, unknowns = parser.parse_known_args(context.argv())
+        if not args.quiet:
+            print 'arguments: ', args
+            print 'unknowns: ', unknowns
+        
+        
+        
+        # Create QWidget
+        self._widget = QWidget()
+        # Get path to UI file which is a sibling of this file
+        # in this example the .ui and .py file are in the same folder
+        ui_file = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'choose_gazebo.ui')
+        # Extend the widget with all attributes and children from UI file
+        loadUi(ui_file, self._widget)
+        # Give QObjects reasonable names
+        self._widget.setObjectName('ChooseGazeboUi')
+        # Show _widget.windowTitle on left-top of each plugin (when 
+        # it's set in _widget). This is useful when you open multiple 
+        # plugins at once. Also if you open multiple instances of your 
+        # plugin at once, these lines add number to make it easy to 
+        # tell from pane to pane.
+        if context.serial_number() > 1:
+            self._widget.setWindowTitle(self._widget.windowTitle() + (' (%d)' % context.serial_number()))
+        # Add widget to the user interface
+        context.add_widget(self._widget)
+
+
+        # Button to start gazebo
+        self._widget.start_gazebo.clicked.connect(self.start_gazebo)
+
+        # Button to kill gazebo
+        self._widget.kill_gazebo.clicked.connect(self.kill_gazebo)  
+
+
+    #@Slot(bool)
+    def start_gazebo(self):
+        #os.system('roslaunch quad_control mav_hovering_example.launch')
+        #subprocess.call('roslaunch quad_control mav_hovering_example.launch &', shell=True)
+        subprocess.call('roslaunch quad_control mav_with_load_example.launch &', shell=True)
+        # subprocess.call('roslaunch quad_control firefly_example.launch &', shell=True)
+        return 
+
+    #@Slot(bool)
+    def kill_gazebo(self):
+        #os.system('killall gzclient')
+        #os.system('killall gzclient;killall gzserver')
+        subprocess.call('killall gzclient &', shell=True)
+        subprocess.call('killall gzserver &', shell=True)
+        return
+
+
+    def _parse_args(self, argv):
+
+        parser = argparse.ArgumentParser(prog='saver', add_help=False)
+
+        # args = parser.parse_args(argv)
+
+        if argv:
+            namespace = argv[0]
+            return namespace
+        else:
+            # is argv is empty return empty string
+            return ""
+    
+    def shutdown_plugin(self):
+        # TODO unregister all publishers here
+        pass
+
+    def save_settings(self, plugin_settings, instance_settings):
+        # TODO save intrinsic configuration, usually using:
+        # instance_settings.set_value(k, v)
+        pass
+
+    def restore_settings(self, plugin_settings, instance_settings):
+        # TODO restore intrinsic configuration, usually using:
+        # v = instance_settings.value(k)
+        pass
+
+    #def trigger_configuration(self):
+        # Comment in to signal that the plugin has a way to configure
+        # This will enable a setting button (gear icon) in each dock widget title bar
+        # Usually used to open a modal configuration dialog
