+++ quad_control/launch/missions_rviz.launch
@@ -1,0 +1,86 @@
+<launch>
+
+    <!-- Controller Node -->
+    <node pkg="quad_control" name="plot_nodeasas" type="plot_experiment.py"  output="screen"/>
+
+    <!-- Default Mission -->
+    <param name="MissionDefault" type="str" value="IrisSimulatorTrajectoryTracking"/>
+
+    <!-- Default Yaw Controller -->
+    <param name="YawControllerDefault" type="str" value="SimpleTrackingYawController"/>
+
+    <!-- Default Trajectory -->
+    <param name="TrajectoryDefault" type="str" value="StayAtRest"/>
+    <param name="trajectry_offset_x" type="double" value="0.0"/>
+    <param name="trajectry_offset_y" type="double" value="0.0"/>
+    <param name="trajectry_offset_z" type="double" value="0.0"/>
+
+
+    <!-- Default Yaw Controller -->
+    <param name="ControllerDefault" type="str" value="SimplePIDController"/>
+    <param name="DIControllerDefault" type="str" value="ComponentWise3DDIC"/>
+
+
+    <!-- Default Vector Thrust Controller -->
+    <param name="VThrustControllerDefault" type="str" value="VThrustQuadrupleController"/>
+
+    <!-- Default Quadruple Integrator Controller -->
+    <param name="QuadrupleIntegratorController" type="str" value="LinearQuadrupleIntController"/>
+
+    <!-- Default Controller for Load transportation by single aerial vehicle -->
+    <param name="SingleLoadTransportationControllerDefault" type="str" value="SingleLoadTransportController"/>
+
+    <!-- Controller Node -->
+    <node pkg="quad_control" name="controller_Iris1" type="cycle_quad_control_mission.py" output="screen"/>
+
+    <!-- Controller parameters -->
+    <param name="gravity_ctr" value="9.81"/>
+    <param name="mass_quad_ctr" value="1.442"/>
+    <param name="Throttle_neutral_ctr" value="1484.0"/>  
+    <param name="ACRO_RP_P_ctr" value="4.5"/> 
+    <param name="ktt" value="20.0"/> 
+    <param name="ktt2" value="30.0"/>
+    <!--  SOME USEFUL RULES: kp = wn*wn  AND kv = 2 xsi wn--> 
+    <param name="kv" value="1.0"/> 
+    <param name="kp" value="1.4"/> 
+    <param name="sigma_p" value="0.5"/> 
+    <param name="sigma_v" value="0.5"/> 
+
+    <param name="throttle_neutral" value="1450"/> 
+    <!-- <param name="quadrotor_mass" value="1.442"/> -->
+    <param name="quadrotor_mass" value="1.669"/>
+
+    <!-- load lifting: cable length in meters -->
+    <param name="cable_length" value="0.6"/> 
+    <!-- load lifting: load mass in kg -->
+    <param name="load_mass" value="0.1"/>
+
+    <!-- Simulator Node -->
+    <node pkg="quad_control" name="Simulator_Iris1" type="quad_simulator.py" output="screen"/>
+
+    <!-- Simulator parameters -->
+    <param name="gravity_sim" value="9.81"/>
+    <param name="mass_quad_sim" value="1.442"/>
+    <param name="Throttle_neutral_sim" value="1484.0"/>  
+    <param name="ACRO_RP_P_sim" value="4.5"/> 
+    <param name="ktt_inner_loop" value="10.0"/> 
+
+
+    <!-- PARAMETERS THAT ARE COMMOM BETWEEN CONTROLLER AND SIMULATOR -->
+    <!-- see WEB: ARDUCOPTER DIFFERENT MODES PARAMETERS -->
+    <param name="MAX_ANGLE_DEG" value="45.0"/> 
+    <param name="MAX_PSI_SPEED_Deg" value="200.0"/> 
+
+
+    <!-- GUI -->
+    <!-- <node name="rqt_gui" pkg="gui" type="tabbedGUI"/> -->
+    <!-- <node name="rqt_gui" pkg="gui" type="tabPlugin"/> -->
+
+    <!-- This runs rqt and we have to select plugin manually -->
+    <!-- similar to rosrun rqt_gui rqt_gui -->
+    <!-- <node name="rqt_gui" pkg="rqt_gui" type="rqt_gui"/> -->
+    <node name="rqt_gui" pkg="rqt_gui" type="rqt_gui" args="--standalone  tabbedGUI --args '' 'rviz' " output="screen"/>
+
+    <node name="rviz" pkg="rviz" type="rviz" args="-d $(find quad_control)/launch/rviz_config.rviz"/>
+
+</launch>
