+++ quad_control/src/missions/mocap/missions_database.py
@@ -1,0 +1,14 @@
+
+#!/usr/bin/env python
+"""
+Database of the missions.
+"""
+
+import rospy
+
+database = {}
+
+import iris_real_trajectory_tracking.iris_real_trajectory_tracking
+database["IrisRealTrajectoryTracking"] = iris_real_trajectory_tracking.iris_real_trajectory_tracking.IrisRealTrajectoryTracking
+
+database["Default"] = database[rospy.get_param("MissionDefault","IrisRealTrajectoryTracking")]
