+++ quad_control/src/missions/rviz/iris_simulator_trajectory_tracking/iris_simulator_trajectory_tracking.py
@@ -1,14 +1,13 @@
 #!/usr/bin/env python
 
 # import Misson abstract class
-from .. import mission
-
-# 
-from mavros_msgs.msg import OverrideRCIn
-# from mavros_msgs.srv import ParamSet,ParamGet,CommandBool,SetMode
+from ... import mission
 
 # import converter from 3d_force and yaw rate into iris rc standard 
 from converter_between_standards.iris_plus_converter import IrisPlusConverter
+
+# publish message quad_cmd that contains commands to simulator
+from quad_control.msg import quad_cmd, quad_state
 
 # import list of available trajectories
 from trajectories import trajectories_database
@@ -25,14 +24,13 @@ import numpy
 # for subscribing to topics, and publishing
 import rospy
 
-import utilities.mocap_source as mocap_source
 
-from utilities.utility_functions import Velocity_Filter
 
-class IrisRealTrajectoryTracking(mission.Mission):
+class IrisSimulatorTrajectoryTracking(mission.Mission):
 
     inner = {}
 
+    
     inner['controller']     = fa_trajectory_tracking_controllers_database.database
     inner['reference']      = trajectories_database.database
     inner['yaw_controller'] = yaw_controllers_database.database
@@ -40,11 +38,10 @@ class IrisRealTrajectoryTracking(mission.Mission):
 
     @classmethod
     def description(cls):
-        return "Firefly, from RotorS, to track a desired trajectory"
-    
+        return "Iris, simulated, to track a desired trajectory"
+
     
     def __init__(self,
-            body_id        = 29,
             controller     = fa_trajectory_tracking_controllers_database.database["Default"](),
             reference      = trajectories_database.database["Default"](),
             yaw_controller = yaw_controllers_database.database["Default"]()
@@ -52,23 +49,14 @@ class IrisRealTrajectoryTracking(mission.Mission):
         # Copy the parameters into self variables
         # Subscribe to the necessary topics, if any
 
-        mission.Mission.__init__(self)
+        mission.Mission.__init__(self)        
+
+        # controller needs to have access to STATE: comes from simulator
+        self.SubToSim = rospy.Subscriber("quad_state", quad_state, self.get_state_from_simulator)
+
+        # message published by quad_control that simulator will subscribe to 
+        self.pub_cmd = rospy.Publisher('quad_cmd', quad_cmd, queue_size=10)
         
-        # converting our controlller standard into iris+ standard
-        self.IrisPlusConverterObject = IrisPlusConverter()
-
-        # establish connection to qualisys
-        self.Qs = mocap_source.Mocap(info=0)
-
-        # publisher: command firefly motor speeds    
-        self.rc_override = rospy.Publisher('mavros/rc/override', OverrideRCIn, queue_size=100)
-
-        self.body_id = body_id
-
-        # intiialization should be done in another way,
-        # but median will take care of minimizing effects
-        self.VelocityEstimator = Velocity_Filter(3,numpy.zeros(3),0.0)
-
         # dy default, desired reference is staying still in origin
         self.TrajGenerator = reference
         self.reference     = self.TrajGenerator.output(self.time_instant_t0)
@@ -78,8 +66,12 @@ class IrisRealTrajectoryTracking(mission.Mission):
 
         self.YawControllerObject = yaw_controller
 
-        self.iris_plus_converter_object_mission.set_mass(self.ControllerObject.MASS)
+        # # converting our controlller standard into iris+ standard
+        # self.IrisPlusConverterObject = IrisPlusConverter()
+        # self.IrisPlusConverterObject.set_mass(self.ControllerObject.MASS)
 
+        self.iris_plus_converter_object_mission.set_mass(self.ControllerObject.MASS)
+        
         pass
 
 
@@ -87,17 +79,17 @@ class IrisRealTrajectoryTracking(mission.Mission):
         # state of quad: position, velocity and attitude 
         # ROLL, PITCH, AND YAW (EULER ANGLES IN DEGREES)
         self.state_quad = numpy.zeros(3+3+3) 
-        pass
-
+        
         
     def __str__(self):
         return self.description()
         # Add the self variables
         
-        
+         
     def __del__(self):
         # Unsubscribe from all the subscribed topics
-        pass
+        self.sub_odometry.unregister()
+
 
     def get_quad_ea_rad(self):
     	euler_rad     = self.state_quad[6:9]*math.pi/180
@@ -112,32 +104,8 @@ class IrisRealTrajectoryTracking(mission.Mission):
 
 
     def get_state(self):
-
-        bodies = self.Qs.get_body(self.body_id)
-
-        if bodies != 'off':  
-
-            self.flag_measurements = True
-
-            # position
-            x=bodies["x"]; y=bodies["y"]; z=bodies["z"]   
-            p = numpy.array([x,y,z])
-
-            # velocity
-            #v = numpy.array([data.vx,data.vy,data.vz])
-            v = self.VelocityEstimator.out(p,rospy.get_time())
-
-            # attitude: euler angles THESE COME IN DEGREES
-            roll = bodies["roll"]; pitch=-bodies["pitch"]; yaw  = bodies["yaw"]
-            ee = numpy.array([roll,pitch,yaw])
-
-            # collect all components of state
-            self.state_quad = numpy.concatenate([p,v,ee])  
-        else:
-            # do nothing, keep previous state
-            self.flag_measurements = False
-
         return self.state_quad
+
 
     def get_pv(self):
         return self.state_quad[0:6]
@@ -147,16 +115,46 @@ class IrisRealTrajectoryTracking(mission.Mission):
         return self.reference[0:6]
 
 
+    def get_complementary_data(self):
+        return numpy.array([])
+
+    def get_labels_complementary_data(self):
+        return []
+
+
     def get_euler_angles(self):
         return self.state_quad[6:9]
 
+
     def real_publish(self,desired_3d_force_quad,yaw_rate,rc_output):
 
-        # ORDER OF INPUTS IS VERY IMPORTANT: roll, pitch, throttle,yaw_rate
-        # create message of type OverrideRCIn
-        rc_cmd          = OverrideRCIn()
-        other_channels  = numpy.array([1500,1500,1500,1500])
-        channels        = numpy.concatenate([rc_output,other_channels])
-        channels        = numpy.array(channels,dtype=numpy.uint16)
-        rc_cmd.channels = channels
-        self.rc_override.publish(rc_cmd)     
+        # create a message of the type quad_cmd, that the simulator subscribes to 
+        cmd  = quad_cmd()
+        
+        cmd.cmd_1 = rc_output[0]
+        cmd.cmd_2 = rc_output[1]
+        cmd.cmd_3 = rc_output[2]
+        cmd.cmd_4 = rc_output[3]
+
+        cmd.cmd_5 = 1500.0
+        cmd.cmd_6 = 1500.0
+        cmd.cmd_7 = 1500.0
+        cmd.cmd_8 = 1500.0
+        
+        self.pub_cmd.publish(cmd)
+
+
+    # callback when simulator publishes states
+    def get_state_from_simulator(self,simulator_message):
+
+        # position
+        p = numpy.array([simulator_message.x,simulator_message.y,simulator_message.z])
+        # velocity
+        v = numpy.array([simulator_message.vx,simulator_message.vy,simulator_message.vz])
+        #v = self.VelocityEstimator.out(p,rospy.get_time())
+        # attitude: euler angles
+        ee = numpy.array([simulator_message.roll,simulator_message.pitch,simulator_message.yaw])
+        # collect all components of state
+        self.state_quad = numpy.concatenate([p,v,ee])  
+
+
