+++ quad_control/src/missions/rviz/missions_database.py
@@ -1,0 +1,14 @@
+
+#!/usr/bin/env python
+"""
+Database of the missions.
+"""
+
+import rospy
+
+database = {}
+
+import iris_simulator_trajectory_tracking.iris_simulator_trajectory_tracking
+database["IrisSimulatorTrajectoryTracking"] = iris_simulator_trajectory_tracking.iris_simulator_trajectory_tracking.IrisSimulatorTrajectoryTracking
+
+database["Default"] = database[rospy.get_param("MissionDefault","IrisSimulatorTrajectoryTracking")]
